rename _app_db
