+++ dbos/dbos.py
@@ -78,7 +78,7 @@ from dbos.context import (
 )
 from dbos.error import DBOSException, DBOSNonExistentWorkflowError
 
-from .application_database import ApplicationDatabase
+from ._app_db import ApplicationDatabase
 from .dbos_config import ConfigFile, load_config, set_env_vars
 from .logger import add_otlp_to_all_loggers, config_logger, dbos_logger, init_logger
 from .system_database import SystemDatabase
